refactor: build images in CI instead of in this project
This removes the Dockerfiles for images I need to customize for the
media stack. Instead of managing the images in this repo, I've migrated
their code to separate projects where CI actions build, tag and push
the images to an OCI registry:

- stackdrift-images/homepage
- stackdrift-images/keycloak
- stackdrift-images/openldap
- stackdrift-images/newt

Now, the compose files can reference those images based on my tags.